backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
pillow
trimesh
scipy
opencv-python-headless
diffusers
transformers
accelerate
safetensors
pydantic
einops
git+https://github.com/openai/shap-e.git
rembg
onnxruntime
ipywidgets

# Note: pytorch should be installed manually for GPU support (see INSTALL.md)
# pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu128
